Rebase and move all changes to PF components. Reset base components
