+++ static/app/views/issueList/issueViews/issueViewQueryCount.tsx
@@ -6,25 +6,25 @@ import {space} from 'sentry/styles/space';
 import type {PageFilters} from 'sentry/types/core';
 import {getUtcDateString} from 'sentry/utils/dates';
 import useOrganization from 'sentry/utils/useOrganization';
+import usePageFilters from 'sentry/utils/usePageFilters';
 import type {IssueView} from 'sentry/views/issueList/issueViews/issueViews';
 import {useFetchIssueCounts} from 'sentry/views/issueList/queries/useFetchIssueCounts';
 
 const TAB_MAX_COUNT = 99;
 
 const constructCountTimeFrame = (
-  timeFilters: PageFilters['datetime']
+  pageFilters: PageFilters['datetime']
 ): {
   end?: string;
   start?: string;
   statsPeriod?: string;
 } => {
-  if (timeFilters.period) {
-    return {statsPeriod: timeFilters.period};
+  if (pageFilters.period) {
+    return {statsPeriod: pageFilters.period};
   }
   return {
-    ...(timeFilters.start ? {start: getUtcDateString(timeFilters.start)} : {}),
-    ...(timeFilters.end ? {end: getUtcDateString(timeFilters.end)} : {}),
-    ...(timeFilters.utc ? {utc: timeFilters.utc} : {}),
+    ...(pageFilters.start ? {start: getUtcDateString(pageFilters.start)} : {}),
+    ...(pageFilters.end ? {end: getUtcDateString(pageFilters.end)} : {}),
   };
 };
 
@@ -34,6 +34,7 @@ interface IssueViewQueryCountProps {
 
 export function IssueViewQueryCount({view}: IssueViewQueryCountProps) {
   const organization = useOrganization();
+  const pageFilters = usePageFilters();
   const theme = useTheme();
 
   // TODO(msun): Once page filters are saved to views, remember to use the view's specific
@@ -46,9 +47,9 @@ export function IssueViewQueryCount({view}: IssueViewQueryCountProps) {
   } = useFetchIssueCounts({
     orgSlug: organization.slug,
     query: [view.unsavedChanges?.query ?? view.query],
-    project: view.unsavedChanges?.projects ?? view.projects,
-    environment: view.unsavedChanges?.environments ?? view.environments,
-    ...constructCountTimeFrame(view.unsavedChanges?.timeFilters ?? view.timeFilters),
+    project: pageFilters.selection.projects,
+    environment: pageFilters.selection.environments,
+    ...constructCountTimeFrame(pageFilters.selection.datetime),
   });
 
   // The endpoint's response type looks like this: { <query1>: <count>, <query2>: <count> }
