+++ static/app/views/issueList/issueViewsHeader.tsx
@@ -1,12 +1,14 @@
-import {useContext, useEffect, useState} from 'react';
+import {useContext, useEffect, useMemo, useState} from 'react';
 import styled from '@emotion/styled';
 import type {Node} from '@react-types/shared';
-import isEqual from 'lodash/isEqual';
 
 import {addSuccessMessage} from 'sentry/actionCreators/indicator';
 import {Button} from 'sentry/components/button';
 import ButtonBar from 'sentry/components/buttonBar';
-import {DraggableTabList} from 'sentry/components/draggableTabs/draggableTabList';
+import {
+  DraggableTabList,
+  TEMPORARY_TAB_KEY,
+} from 'sentry/components/draggableTabs/draggableTabList';
 import type {DraggableTabListItemProps} from 'sentry/components/draggableTabs/item';
 import GlobalEventProcessingAlert from 'sentry/components/globalEventProcessingAlert';
 import * as Layout from 'sentry/components/layouts/thirds';
@@ -23,16 +25,15 @@ import {useHotkeys} from 'sentry/utils/useHotkeys';
 import {useLocation} from 'sentry/utils/useLocation';
 import {useNavigate} from 'sentry/utils/useNavigate';
 import useOrganization from 'sentry/utils/useOrganization';
+import usePageFilters from 'sentry/utils/usePageFilters';
 import useProjects from 'sentry/utils/useProjects';
+import {useUser} from 'sentry/utils/useUser';
 import {
-  DEFAULT_ENVIRONMENTS,
-  DEFAULT_TIME_FILTERS,
   generateTempViewId,
   type IssueView,
   type IssueViewParams,
   IssueViews,
   IssueViewsContext,
-  TEMPORARY_TAB_KEY,
 } from 'sentry/views/issueList/issueViews/issueViews';
 import {IssueViewTab} from 'sentry/views/issueList/issueViews/issueViewTab';
 import {useFetchGroupSearchViews} from 'sentry/views/issueList/queries/useFetchGroupSearchViews';
@@ -135,16 +136,7 @@ function IssueViewsIssueListHeader({
           router={router}
           initialViews={groupSearchViews.map(
             (
-              {
-                id,
-                name,
-                query: viewQuery,
-                querySort: viewQuerySort,
-                environments: viewEnvironments,
-                projects: viewProjects,
-                timeFilters: viewTimeFilters,
-                isAllProjects,
-              },
+              {id, name, query: viewQuery, querySort: viewQuerySort},
               index
             ): IssueView => {
               const tabId = id ?? `default${index.toString()}`;
@@ -155,9 +147,6 @@ function IssueViewsIssueListHeader({
                 label: name,
                 query: viewQuery,
                 querySort: viewQuerySort,
-                environments: viewEnvironments,
-                projects: isAllProjects ? [-1] : viewProjects,
-                timeFilters: viewTimeFilters,
                 isCommitted: true,
               };
             }
@@ -178,48 +167,58 @@ function IssueViewsIssueListHeaderTabsContent({
   const organization = useOrganization();
   const navigate = useNavigate();
   const location = useLocation();
+  const pageFilters = usePageFilters();
+  const user = useUser();
 
   const {newViewActive, setNewViewActive} = useContext(NewTabContext);
-  const {tabListState, state, dispatch, defaultProject} = useContext(IssueViewsContext);
+  const {tabListState, state, dispatch} = useContext(IssueViewsContext);
   const {views, tempView} = state;
 
   const [editingTabKey, setEditingTabKey] = useState<string | null>(null);
 
   // TODO(msun): Use the location from useLocation instead of props router in the future
-  const {query, sort, viewId} = router.location.query;
+  const queryParams = router.location.query;
+  const {query, sort, viewId, project, environment} = queryParams;
+  const queryParamsWithPageFilters = useMemo(() => {
+    return {
+      ...queryParams,
+      project: project ?? pageFilters.selection.projects,
+      environment: environment ?? pageFilters.selection.environments,
+      ...normalizeDateTimeParams(pageFilters.selection.datetime),
+    };
+  }, [
+    environment,
+    pageFilters.selection.datetime,
+    pageFilters.selection.environments,
+    pageFilters.selection.projects,
+    project,
+    queryParams,
+  ]);
+
+  // This useEffect is here temporarily to start saving user's most recently used page filters
+  // to their views. This will be removed once the frontend is updated to use a view's page filters
+  useEffect(() => {
+    dispatch({type: 'SYNC_VIEWS_TO_BACKEND'});
+
+    trackAnalytics('issue_views.page_filters_logged', {
+      user_id: user.id,
+      organization,
+    });
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [pageFilters.selection]);
 
   // This insane useEffect ensures that the correct tab is selected when the url updates
   useEffect(() => {
-    const {
-      project,
-      environment: env,
-      start,
-      end,
-      statsPeriod,
-      utc,
-    } = router.location.query;
     // If no query, sort, or viewId is present, set the first tab as the selected tab, update query accordingly
     if (!query && !sort && !viewId) {
-      const {
-        id,
-        query: viewQuery,
-        querySort,
-        projects,
-        environments,
-        timeFilters,
-      } = views[0]!;
-
       navigate(
         normalizeUrl({
           ...location,
           query: {
-            ...router.location.query,
-            query: viewQuery,
-            sort: querySort,
-            viewId: id,
-            project: projects,
-            environment: environments,
-            ...normalizeDateTimeParams(timeFilters),
+            ...queryParamsWithPageFilters,
+            query: views[0]!.query,
+            sort: views[0]!.querySort,
+            viewId: views[0]!.id,
           },
         }),
         {replace: true}
@@ -227,65 +226,37 @@ function IssueViewsIssueListHeaderTabsContent({
       tabListState?.setSelectedKey(views[0]!.key);
       return;
     }
-    const {queryEnvs, queryProjects} = normalizeProjectsEnvironments(project, env);
-    const queryTimeFilters =
-      start || end || statsPeriod || utc
-        ? {
-            start: statsPeriod ? null : start,
-            end: statsPeriod ? null : end,
-            period: statsPeriod,
-            utc: statsPeriod ? null : utc,
-          }
-        : undefined;
     // if a viewId is present, check the frontend is aware of a view with that id.
     if (viewId) {
       const selectedView = views.find(tab => tab.id === viewId);
       if (selectedView) {
         // If the frontend is aware of a view with that id, check if the query/sort is different
         // from the view's original query/sort.
-        const {
-          query: originalQuery,
-          querySort: originalSort,
-          projects: originalProjects,
-          environments: originalEnvironments,
-          timeFilters: originalTimeFilters,
-        } = selectedView;
-
+        const {query: originalQuery, querySort: originalSort} = selectedView;
         const issueSortOption = Object.values(IssueSortOptions).includes(sort)
-          ? (sort as IssueSortOptions)
-          : undefined;
+          ? sort
+          : IssueSortOptions.DATE;
 
-        const newUnsavedChanges: Partial<IssueViewParams> = {
-          query: query === originalQuery ? undefined : query,
-          querySort: sort === originalSort ? undefined : issueSortOption,
-          projects: isEqual(queryProjects.sort(), originalProjects.sort())
+        const newUnsavedChanges: IssueViewParams | undefined =
+          query === originalQuery && sort === originalSort
             ? undefined
-            : queryProjects,
-          environments: isEqual(queryEnvs.sort(), originalEnvironments.sort())
-            ? undefined
-            : queryEnvs,
-          timeFilters:
-            queryTimeFilters &&
-            isEqual(
-              normalizeDateTimeParams(originalTimeFilters),
-              normalizeDateTimeParams(queryTimeFilters)
-            )
-              ? undefined
-              : queryTimeFilters,
-        };
-        const hasNoChanges = Object.values(newUnsavedChanges).every(
-          value => value === undefined
-        );
-
-        if (!hasNoChanges && !isEqual(selectedView.unsavedChanges, newUnsavedChanges)) {
+            : {
+                query,
+                querySort: issueSortOption,
+              };
+        if (
+          (newUnsavedChanges && !selectedView.unsavedChanges) ||
+          selectedView.unsavedChanges?.query !== newUnsavedChanges?.query ||
+          selectedView.unsavedChanges?.querySort !== newUnsavedChanges?.querySort
+        ) {
           // If there were no unsaved changes before, or the existing unsaved changes
           // don't match the new query and/or sort, update the unsaved changes
           dispatch({
             type: 'UPDATE_UNSAVED_CHANGES',
             unsavedChanges: newUnsavedChanges,
           });
-        } else if (hasNoChanges && selectedView.unsavedChanges) {
-          // If all view params are the same as the original view params, remove any unsaved changes
+        } else if (!newUnsavedChanges && selectedView.unsavedChanges) {
+          // If the query/sort are the same as the original query/sort, remove any unsaved changes
           dispatch({type: 'UPDATE_UNSAVED_CHANGES', unsavedChanges: undefined});
         }
         if (!tabListState?.selectionManager.isSelected(selectedView.key)) {
@@ -293,15 +264,10 @@ function IssueViewsIssueListHeaderTabsContent({
             normalizeUrl({
               ...location,
               query: {
-                ...router.location.query,
+                ...queryParamsWithPageFilters,
+                query: newUnsavedChanges?.query ?? selectedView.query,
+                sort: newUnsavedChanges?.querySort ?? selectedView.querySort,
                 viewId: selectedView.id,
-                query: newUnsavedChanges.query ?? selectedView.query,
-                sort: newUnsavedChanges.querySort ?? selectedView.querySort,
-                project: newUnsavedChanges.projects ?? selectedView.projects,
-                environment: newUnsavedChanges.environments ?? selectedView.environments,
-                ...normalizeDateTimeParams(
-                  newUnsavedChanges.timeFilters ?? selectedView.timeFilters
-                ),
               },
             }),
             {replace: true}
@@ -309,17 +275,14 @@ function IssueViewsIssueListHeaderTabsContent({
           tabListState?.setSelectedKey(selectedView.key);
         }
       } else {
-        // if the viewId isn't found in this user's views, remove it from the query
+        // if a viewId does not exist, remove it from the query
         tabListState?.setSelectedKey(TEMPORARY_TAB_KEY);
         navigate(
           normalizeUrl({
             ...location,
             query: {
-              ...router.location.query,
+              ...queryParamsWithPageFilters,
               viewId: undefined,
-              project: project ?? defaultProject,
-              environment: env ?? DEFAULT_ENVIRONMENTS,
-              ...normalizeDateTimeParams(queryTimeFilters ?? DEFAULT_TIME_FILTERS),
             },
           }),
           {replace: true}
@@ -338,7 +301,7 @@ function IssueViewsIssueListHeaderTabsContent({
           normalizeUrl({
             ...location,
             query: {
-              ...router.location.query,
+              ...queryParamsWithPageFilters,
               viewId: undefined,
             },
           }),
@@ -349,17 +312,17 @@ function IssueViewsIssueListHeaderTabsContent({
       }
     }
   }, [
-    defaultProject,
-    dispatch,
-    location,
     navigate,
-    organization,
+    organization.slug,
     query,
-    router.location.query,
     sort,
-    tabListState,
     viewId,
+    tabListState,
+    location,
+    queryParamsWithPageFilters,
     views,
+    organization,
+    dispatch,
   ]);
 
   // This useEffect ensures the "new view" page is displayed/hidden correctly
@@ -415,12 +378,9 @@ function IssueViewsIssueListHeaderTabsContent({
     navigate({
       ...location,
       query: {
-        ...router.location.query,
+        ...queryParams,
         query: '',
         viewId: tempId,
-        project: defaultProject,
-        environment: DEFAULT_ENVIRONMENTS,
-        ...normalizeDateTimeParams(DEFAULT_TIME_FILTERS),
       },
     });
   };
@@ -455,15 +415,10 @@ function IssueViewsIssueListHeaderTabsContent({
           key={view.key}
           to={normalizeUrl({
             query: {
-              ...router.location.query,
+              ...queryParams,
               query: view.unsavedChanges?.query ?? view.query,
               sort: view.unsavedChanges?.querySort ?? view.querySort,
               viewId: view.id !== TEMPORARY_TAB_KEY ? view.id : undefined,
-              project: view.unsavedChanges?.projects ?? view.projects,
-              environment: view.unsavedChanges?.environments ?? view.environments,
-              ...normalizeDateTimeParams(
-                view.unsavedChanges?.timeFilters ?? view.timeFilters
-              ),
               cursor: undefined,
               page: undefined,
             },
@@ -486,35 +441,6 @@ function IssueViewsIssueListHeaderTabsContent({
 }
 
 export default IssueViewsIssueListHeader;
-
-/**
- * Normalizes the project and environment query params to arrays of strings and numbers.
- * If project/environemnts is undefined, it equates to an empty array. If it is a single value,
- * it is converted to single element array.
- */
-const normalizeProjectsEnvironments = (
-  project: any[] | undefined,
-  env: any[] | undefined
-): {queryEnvs: string[]; queryProjects: number[]} => {
-  let queryProjects: number[] = [];
-  if (Array.isArray(project)) {
-    queryProjects = project.map(p => parseInt(p, 10)).filter(p => !isNaN(p));
-  } else if (project) {
-    const parsed = parseInt(project, 10);
-    if (!isNaN(parsed)) {
-      queryProjects = [parsed];
-    }
-  }
-
-  let queryEnvs: string[] = [];
-  if (Array.isArray(env)) {
-    queryEnvs = env;
-  } else if (env) {
-    queryEnvs = [env];
-  }
-
-  return {queryEnvs, queryProjects};
-};
 
 const StyledIssueViews = styled(IssueViews)`
   grid-column: 1 / -1;
