+++ static/app/views/issueList/issueViewsHeaderPF.tsx
@@ -1,11 +1,11 @@
-import {useContext, useEffect, useMemo, useState} from 'react';
+import {useContext, useEffect, useState} from 'react';
 import styled from '@emotion/styled';
 import type {Node} from '@react-types/shared';
+import isEqual from 'lodash/isEqual';
 
 import {addSuccessMessage} from 'sentry/actionCreators/indicator';
 import {Button} from 'sentry/components/button';
 import ButtonBar from 'sentry/components/buttonBar';
-import {TEMPORARY_TAB_KEY} from 'sentry/components/draggableTabs/draggableTabList';
 import {DraggableTabListPF} from 'sentry/components/draggableTabs/draggableTabListPF';
 import type {DraggableTabListItemProps} from 'sentry/components/draggableTabs/item';
 import GlobalEventProcessingAlert from 'sentry/components/globalEventProcessingAlert';
@@ -23,15 +23,16 @@ import {useHotkeys} from 'sentry/utils/useHotkeys';
 import {useLocation} from 'sentry/utils/useLocation';
 import {useNavigate} from 'sentry/utils/useNavigate';
 import useOrganization from 'sentry/utils/useOrganization';
-import usePageFilters from 'sentry/utils/usePageFilters';
 import useProjects from 'sentry/utils/useProjects';
-import {useUser} from 'sentry/utils/useUser';
 import {
+  DEFAULT_ENVIRONMENTS,
+  DEFAULT_TIME_FILTERS,
   generateTempViewId,
   type IssueViewPF,
   type IssueViewPFParams,
   IssueViewsPF,
   IssueViewsPFContext,
+  TEMPORARY_TAB_KEY,
 } from 'sentry/views/issueList/issueViewsPF/issueViewsPF';
 import {IssueViewPFTab} from 'sentry/views/issueList/issueViewsPF/issueViewTabPF';
 import {useFetchGroupSearchViews} from 'sentry/views/issueList/queries/useFetchGroupSearchViews';
@@ -39,7 +40,7 @@ import {NewTabContext} from 'sentry/views/issueList/utils/newTabContext';
 
 import {IssueSortOptions} from './utils';
 
-type IssueViewsIssueListHeaderProps = {
+type IssueViewsPFIssueListHeaderProps = {
   onRealtimeChange: (realtime: boolean) => void;
   realtimeActive: boolean;
   router: InjectedRouter;
@@ -55,7 +56,7 @@ function IssueViewsPFIssueListHeader({
   realtimeActive,
   onRealtimeChange,
   router,
-}: IssueViewsIssueListHeaderProps) {
+}: IssueViewsPFIssueListHeaderProps) {
   const organization = useOrganization();
   const {projects} = useProjects();
   const selectedProjects = projects.filter(({id}) =>
@@ -134,7 +135,16 @@ function IssueViewsPFIssueListHeader({
           router={router}
           initialViews={groupSearchViews.map(
             (
-              {id, name, query: viewQuery, querySort: viewQuerySort},
+              {
+                id,
+                name,
+                query: viewQuery,
+                querySort: viewQuerySort,
+                environments: viewEnvironments,
+                projects: viewProjects,
+                timeFilters: viewTimeFilters,
+                isAllProjects,
+              },
               index
             ): IssueViewPF => {
               const tabId = id ?? `default${index.toString()}`;
@@ -145,12 +155,15 @@ function IssueViewsPFIssueListHeader({
                 label: name,
                 query: viewQuery,
                 querySort: viewQuerySort,
+                environments: viewEnvironments,
+                projects: isAllProjects ? [-1] : viewProjects,
+                timeFilters: viewTimeFilters,
                 isCommitted: true,
               };
             }
           )}
         >
-          <IssueViewsPFIssueListHeaderTabsContent router={router} />
+          <IssueViewsIssueListHeaderTabsContent router={router} />
         </StyledIssueViews>
       ) : (
         <div style={{height: 33}} />
@@ -159,64 +172,54 @@ function IssueViewsPFIssueListHeader({
   );
 }
 
-function IssueViewsPFIssueListHeaderTabsContent({
+function IssueViewsIssueListHeaderTabsContent({
   router,
 }: IssueViewsIssueListHeaderTabsContentProps) {
   const organization = useOrganization();
   const navigate = useNavigate();
   const location = useLocation();
-  const pageFilters = usePageFilters();
-  const user = useUser();
 
   const {newViewActive, setNewViewActive} = useContext(NewTabContext);
-  const {tabListState, state, dispatch} = useContext(IssueViewsPFContext);
+  const {tabListState, state, dispatch, defaultProject} = useContext(IssueViewsPFContext);
   const {views, tempView} = state;
 
   const [editingTabKey, setEditingTabKey] = useState<string | null>(null);
 
   // TODO(msun): Use the location from useLocation instead of props router in the future
-  const queryParams = router.location.query;
-  const {query, sort, viewId, project, environment} = queryParams;
-  const queryParamsWithPageFilters = useMemo(() => {
-    return {
-      ...queryParams,
-      project: project ?? pageFilters.selection.projects,
-      environment: environment ?? pageFilters.selection.environments,
-      ...normalizeDateTimeParams(pageFilters.selection.datetime),
-    };
-  }, [
-    environment,
-    pageFilters.selection.datetime,
-    pageFilters.selection.environments,
-    pageFilters.selection.projects,
-    project,
-    queryParams,
-  ]);
-
-  // This useEffect is here temporarily to start saving user's most recently used page filters
-  // to their views. This will be removed once the frontend is updated to use a view's page filters
-  useEffect(() => {
-    dispatch({type: 'SYNC_VIEWS_TO_BACKEND'});
-
-    trackAnalytics('issue_views.page_filters_logged', {
-      user_id: user.id,
-      organization,
-    });
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [pageFilters.selection]);
+  const {query, sort, viewId} = router.location.query;
 
   // This insane useEffect ensures that the correct tab is selected when the url updates
   useEffect(() => {
+    const {
+      project,
+      environment: env,
+      start,
+      end,
+      statsPeriod,
+      utc,
+    } = router.location.query;
     // If no query, sort, or viewId is present, set the first tab as the selected tab, update query accordingly
     if (!query && !sort && !viewId) {
+      const {
+        id,
+        query: viewQuery,
+        querySort,
+        projects,
+        environments,
+        timeFilters,
+      } = views[0]!;
+
       navigate(
         normalizeUrl({
           ...location,
           query: {
-            ...queryParamsWithPageFilters,
-            query: views[0]!.query,
-            sort: views[0]!.querySort,
-            viewId: views[0]!.id,
+            ...router.location.query,
+            query: viewQuery,
+            sort: querySort,
+            viewId: id,
+            project: projects,
+            environment: environments,
+            ...normalizeDateTimeParams(timeFilters),
           },
         }),
         {replace: true}
@@ -224,37 +227,65 @@ function IssueViewsPFIssueListHeaderTabsContent({
       tabListState?.setSelectedKey(views[0]!.key);
       return;
     }
+    const {queryEnvs, queryProjects} = normalizeProjectsEnvironments(project, env);
+    const queryTimeFilters =
+      start || end || statsPeriod || utc
+        ? {
+            start: statsPeriod ? null : start,
+            end: statsPeriod ? null : end,
+            period: statsPeriod,
+            utc: statsPeriod ? null : utc,
+          }
+        : undefined;
     // if a viewId is present, check the frontend is aware of a view with that id.
     if (viewId) {
       const selectedView = views.find(tab => tab.id === viewId);
       if (selectedView) {
         // If the frontend is aware of a view with that id, check if the query/sort is different
         // from the view's original query/sort.
-        const {query: originalQuery, querySort: originalSort} = selectedView;
-        const issueSortOption = Object.values(IssueSortOptions).includes(sort)
-          ? sort
-          : IssueSortOptions.DATE;
+        const {
+          query: originalQuery,
+          querySort: originalSort,
+          projects: originalProjects,
+          environments: originalEnvironments,
+          timeFilters: originalTimeFilters,
+        } = selectedView;
 
-        const newUnsavedChanges: IssueViewPFParams | undefined =
-          query === originalQuery && sort === originalSort
+        const issueSortOption = Object.values(IssueSortOptions).includes(sort)
+          ? (sort as IssueSortOptions)
+          : undefined;
+
+        const newUnsavedChanges: Partial<IssueViewPFParams> = {
+          query: query === originalQuery ? undefined : query,
+          querySort: sort === originalSort ? undefined : issueSortOption,
+          projects: isEqual(queryProjects.sort(), originalProjects.sort())
             ? undefined
-            : {
-                query,
-                querySort: issueSortOption,
-              };
-        if (
-          (newUnsavedChanges && !selectedView.unsavedChanges) ||
-          selectedView.unsavedChanges?.query !== newUnsavedChanges?.query ||
-          selectedView.unsavedChanges?.querySort !== newUnsavedChanges?.querySort
-        ) {
+            : queryProjects,
+          environments: isEqual(queryEnvs.sort(), originalEnvironments.sort())
+            ? undefined
+            : queryEnvs,
+          timeFilters:
+            queryTimeFilters &&
+            isEqual(
+              normalizeDateTimeParams(originalTimeFilters),
+              normalizeDateTimeParams(queryTimeFilters)
+            )
+              ? undefined
+              : queryTimeFilters,
+        };
+        const hasNoChanges = Object.values(newUnsavedChanges).every(
+          value => value === undefined
+        );
+
+        if (!hasNoChanges && !isEqual(selectedView.unsavedChanges, newUnsavedChanges)) {
           // If there were no unsaved changes before, or the existing unsaved changes
           // don't match the new query and/or sort, update the unsaved changes
           dispatch({
             type: 'UPDATE_UNSAVED_CHANGES',
             unsavedChanges: newUnsavedChanges,
           });
-        } else if (!newUnsavedChanges && selectedView.unsavedChanges) {
-          // If the query/sort are the same as the original query/sort, remove any unsaved changes
+        } else if (hasNoChanges && selectedView.unsavedChanges) {
+          // If all view params are the same as the original view params, remove any unsaved changes
           dispatch({type: 'UPDATE_UNSAVED_CHANGES', unsavedChanges: undefined});
         }
         if (!tabListState?.selectionManager.isSelected(selectedView.key)) {
@@ -262,10 +293,15 @@ function IssueViewsPFIssueListHeaderTabsContent({
             normalizeUrl({
               ...location,
               query: {
-                ...queryParamsWithPageFilters,
-                query: newUnsavedChanges?.query ?? selectedView.query,
-                sort: newUnsavedChanges?.querySort ?? selectedView.querySort,
+                ...router.location.query,
                 viewId: selectedView.id,
+                query: newUnsavedChanges.query ?? selectedView.query,
+                sort: newUnsavedChanges.querySort ?? selectedView.querySort,
+                project: newUnsavedChanges.projects ?? selectedView.projects,
+                environment: newUnsavedChanges.environments ?? selectedView.environments,
+                ...normalizeDateTimeParams(
+                  newUnsavedChanges.timeFilters ?? selectedView.timeFilters
+                ),
               },
             }),
             {replace: true}
@@ -273,14 +309,17 @@ function IssueViewsPFIssueListHeaderTabsContent({
           tabListState?.setSelectedKey(selectedView.key);
         }
       } else {
-        // if a viewId does not exist, remove it from the query
+        // if the viewId isn't found in this user's views, remove it from the query
         tabListState?.setSelectedKey(TEMPORARY_TAB_KEY);
         navigate(
           normalizeUrl({
             ...location,
             query: {
-              ...queryParamsWithPageFilters,
+              ...router.location.query,
               viewId: undefined,
+              project: project ?? defaultProject,
+              environment: env ?? DEFAULT_ENVIRONMENTS,
+              ...normalizeDateTimeParams(queryTimeFilters ?? DEFAULT_TIME_FILTERS),
             },
           }),
           {replace: true}
@@ -299,7 +338,7 @@ function IssueViewsPFIssueListHeaderTabsContent({
           normalizeUrl({
             ...location,
             query: {
-              ...queryParamsWithPageFilters,
+              ...router.location.query,
               viewId: undefined,
             },
           }),
@@ -310,17 +349,17 @@ function IssueViewsPFIssueListHeaderTabsContent({
       }
     }
   }, [
-    navigate,
-    organization.slug,
-    query,
-    sort,
-    viewId,
-    tabListState,
-    location,
-    queryParamsWithPageFilters,
-    views,
-    organization,
+    defaultProject,
     dispatch,
+    location,
+    navigate,
+    organization,
+    query,
+    router.location.query,
+    sort,
+    tabListState,
+    viewId,
+    views,
   ]);
 
   // This useEffect ensures the "new view" page is displayed/hidden correctly
@@ -376,9 +415,12 @@ function IssueViewsPFIssueListHeaderTabsContent({
     navigate({
       ...location,
       query: {
-        ...queryParams,
+        ...router.location.query,
         query: '',
         viewId: tempId,
+        project: defaultProject,
+        environment: DEFAULT_ENVIRONMENTS,
+        ...normalizeDateTimeParams(DEFAULT_TIME_FILTERS),
       },
     });
   };
@@ -413,10 +455,15 @@ function IssueViewsPFIssueListHeaderTabsContent({
           key={view.key}
           to={normalizeUrl({
             query: {
-              ...queryParams,
+              ...router.location.query,
               query: view.unsavedChanges?.query ?? view.query,
               sort: view.unsavedChanges?.querySort ?? view.querySort,
               viewId: view.id !== TEMPORARY_TAB_KEY ? view.id : undefined,
+              project: view.unsavedChanges?.projects ?? view.projects,
+              environment: view.unsavedChanges?.environments ?? view.environments,
+              ...normalizeDateTimeParams(
+                view.unsavedChanges?.timeFilters ?? view.timeFilters
+              ),
               cursor: undefined,
               page: undefined,
             },
@@ -439,6 +486,35 @@ function IssueViewsPFIssueListHeaderTabsContent({
 }
 
 export default IssueViewsPFIssueListHeader;
+
+/**
+ * Normalizes the project and environment query params to arrays of strings and numbers.
+ * If project/environemnts is undefined, it equates to an empty array. If it is a single value,
+ * it is converted to single element array.
+ */
+const normalizeProjectsEnvironments = (
+  project: any[] | undefined,
+  env: any[] | undefined
+): {queryEnvs: string[]; queryProjects: number[]} => {
+  let queryProjects: number[] = [];
+  if (Array.isArray(project)) {
+    queryProjects = project.map(p => parseInt(p, 10)).filter(p => !isNaN(p));
+  } else if (project) {
+    const parsed = parseInt(project, 10);
+    if (!isNaN(parsed)) {
+      queryProjects = [parsed];
+    }
+  }
+
+  let queryEnvs: string[] = [];
+  if (Array.isArray(env)) {
+    queryEnvs = env;
+  } else if (env) {
+    queryEnvs = [env];
+  }
+
+  return {queryEnvs, queryProjects};
+};
 
 const StyledIssueViews = styled(IssueViewsPF)`
   grid-column: 1 / -1;
