+++ static/app/views/issueList/issueViewsPF/issueViewsPF.tsx
@@ -12,8 +12,8 @@ import styled from '@emotion/styled';
 import type {TabListState} from '@react-stately/tabs';
 import type {Orientation} from '@react-types/shared';
 import debounce from 'lodash/debounce';
+import isEqual from 'lodash/isEqual';
 
-import {normalizeDateTimeParams} from 'sentry/components/organizations/pageFilters/parse';
 import type {TabContext, TabsProps} from 'sentry/components/tabs';
 import {tabsShouldForwardProp} from 'sentry/components/tabs/utils';
 import {t} from 'sentry/locale';
@@ -24,7 +24,7 @@ import {trackAnalytics} from 'sentry/utils/analytics';
 import normalizeUrl from 'sentry/utils/url/normalizeUrl';
 import {useNavigate} from 'sentry/utils/useNavigate';
 import useOrganization from 'sentry/utils/useOrganization';
-import usePageFilters from 'sentry/utils/usePageFilters';
+import useProjects from 'sentry/utils/useProjects';
 import {useUpdateGroupSearchViews} from 'sentry/views/issueList/mutations/useUpdateGroupSearchViews';
 import type {
   GroupSearchView,
@@ -33,13 +33,29 @@ import type {
 import {IssueSortOptions} from 'sentry/views/issueList/utils';
 import {NewTabContext, type NewView} from 'sentry/views/issueList/utils/newTabContext';
 
-const TEMPORARY_TAB_KEY = 'temporary-tab';
+export const TEMPORARY_TAB_KEY = 'temporary-tab';
+
+export const DEFAULT_TIME_FILTERS: PageFilters['datetime'] = {
+  start: null,
+  end: null,
+  period: '14d',
+  utc: null,
+};
+export const DEFAULT_ENVIRONMENTS: string[] = [];
 
 export const generateTempViewId = () => `_${Math.random().toString().substring(2, 7)}`;
 
+/**
+ * Savable properties of an IssueView, besides lable and position.
+ * Changes to these properties are not automatically saved and can
+ * trigger the unsaved changes indicator.
+ */
 export interface IssueViewPFParams {
+  environments: string[];
+  projects: number[];
   query: string;
   querySort: IssueSortOptions;
+  timeFilters: PageFilters['datetime'];
 }
 
 export interface IssueViewPF extends IssueViewPFParams {
@@ -53,7 +69,7 @@ export interface IssueViewPF extends IssueViewPFParams {
   key: string;
   label: string;
   content?: React.ReactNode;
-  unsavedChanges?: IssueViewPFParams;
+  unsavedChanges?: Partial<IssueViewPFParams>;
 }
 
 type BaseIssueViewsAction = {
@@ -110,7 +126,7 @@ type SaveTempViewAction = {
 type UpdateUnsavedChangesAction = {
   type: 'UPDATE_UNSAVED_CHANGES';
   // Explicitly typed as | undefined instead of optional to make it clear that `undefined` = no unsaved changes
-  unsavedChanges: IssueViewPFParams | undefined;
+  unsavedChanges: Partial<IssueViewPFParams> | undefined;
   isCommitted?: boolean;
 } & BaseIssueViewsAction;
 
@@ -129,7 +145,7 @@ type SyncViewsToBackendAction = {
   type: 'SYNC_VIEWS_TO_BACKEND';
 };
 
-export type IssueViewsPFActions =
+export type IssueViewsActions =
   | ReorderTabsAction
   | SaveChangesAction
   | DiscardChangesAction
@@ -145,7 +161,7 @@ export type IssueViewsPFActions =
   | SetViewsAction
   | SyncViewsToBackendAction;
 
-const ACTION_ANALYTICS_MAP: Partial<Record<IssueViewsPFActions['type'], string>> = {
+const ACTION_ANALYTICS_MAP: Partial<Record<IssueViewsActions['type'], string>> = {
   REORDER_TABS: 'issue_views.reordered_views',
   SAVE_CHANGES: 'issue_views.saved_changes',
   DISCARD_CHANGES: 'issue_views.discarded_changes',
@@ -163,7 +179,8 @@ export interface IssueViewsPFState {
 }
 
 export interface IssueViewsPFContextType extends TabContext {
-  dispatch: Dispatch<IssueViewsPFActions>;
+  defaultProject: number[];
+  dispatch: Dispatch<IssueViewsActions>;
   state: IssueViewsPFState;
 }
 
@@ -173,6 +190,7 @@ export const IssueViewsPFContext = createContext<IssueViewsPFContextType>({
   // Issue Views specific state
   dispatch: () => {},
   state: {views: []},
+  defaultProject: [],
 });
 
 function reorderTabs(state: IssueViewsPFState, action: ReorderTabsAction) {
@@ -192,8 +210,11 @@ function saveChanges(state: IssueViewsPFState, tabListState: TabListState<any>) 
       return tab.key === tabListState?.selectedKey && tab.unsavedChanges
         ? {
             ...tab,
-            query: tab.unsavedChanges.query,
-            querySort: tab.unsavedChanges.querySort,
+            query: tab.unsavedChanges.query ?? tab.query,
+            querySort: tab.unsavedChanges.querySort ?? tab.querySort,
+            projects: tab.unsavedChanges.projects ?? tab.projects,
+            environments: tab.unsavedChanges.environments ?? tab.environments,
+            timeFilters: tab.unsavedChanges.timeFilters ?? tab.timeFilters,
             unsavedChanges: undefined,
           }
         : tab;
@@ -262,7 +283,11 @@ function deleteView(state: IssueViewsPFState, tabListState: TabListState<any>) {
   return {...state, views: newViews};
 }
 
-function createNewView(state: IssueViewsPFState, action: CreateNewViewAction) {
+function createNewView(
+  state: IssueViewsPFState,
+  action: CreateNewViewAction,
+  defaultProject: number[]
+) {
   const newTabs: IssueViewPF[] = [
     ...state.views,
     {
@@ -272,12 +297,19 @@ function createNewView(state: IssueViewsPFState, action: CreateNewViewAction) {
       query: '',
       querySort: IssueSortOptions.DATE,
       isCommitted: false,
+      environments: DEFAULT_ENVIRONMENTS,
+      projects: defaultProject,
+      timeFilters: DEFAULT_TIME_FILTERS,
     },
   ];
   return {...state, views: newTabs};
 }
 
-function setTempView(state: IssueViewsPFState, action: SetTempViewAction) {
+function setTempView(
+  state: IssueViewsPFState,
+  action: SetTempViewAction,
+  defaultProject: number[]
+) {
   const tempView: IssueViewPF = {
     id: TEMPORARY_TAB_KEY,
     key: TEMPORARY_TAB_KEY,
@@ -285,6 +317,9 @@ function setTempView(state: IssueViewsPFState, action: SetTempViewAction) {
     query: action.query,
     querySort: action.sort ?? IssueSortOptions.DATE,
     isCommitted: true,
+    environments: DEFAULT_ENVIRONMENTS,
+    projects: defaultProject,
+    timeFilters: DEFAULT_TIME_FILTERS,
   };
   return {...state, tempView};
 }
@@ -304,6 +339,9 @@ function saveTempView(state: IssueViewsPFState, tabListState: TabListState<any>)
       query: state.tempView?.query,
       querySort: state.tempView?.querySort,
       isCommitted: true,
+      environments: state.tempView?.environments,
+      projects: state.tempView?.projects,
+      timeFilters: state.tempView?.timeFilters,
     };
     tabListState?.setSelectedKey(tempId);
     return {...state, views: [...state.views, newTab], tempView: undefined};
@@ -370,30 +408,26 @@ export function IssueViewsPFStateProvider({
   ...props
 }: IssueViewsPFStateProviderProps) {
   const navigate = useNavigate();
-  const pageFilters = usePageFilters();
   const organization = useOrganization();
   const {setNewViewActive, setOnNewViewsSaved} = useContext(NewTabContext);
   const [tabListState, setTabListState] = useState<TabListState<any>>();
   const {className: _className, ...restProps} = props;
 
-  const {cursor: _cursor, page: _page, ...queryParams} = router.location.query;
-  const {query, sort, viewId, project, environment} = queryParams;
+  const allowMultipleProjects = organization.features.includes('global-views');
+  const {projects: allProjects} = useProjects();
+  const memberProjects = useMemo(
+    () => allProjects.filter(project => project.isMember),
+    [allProjects]
+  );
+  const defaultProject = useMemo(() => {
+    // Should not be possible for member projects to be empty
+    if (allowMultipleProjects) {
+      return [];
+    }
+    return [parseInt(memberProjects[0]!.id, 10)];
+  }, [memberProjects, allowMultipleProjects]);
 
-  const queryParamsWithPageFilters = useMemo(() => {
-    return {
-      ...queryParams,
-      project: project ?? pageFilters.selection.projects,
-      environment: environment ?? pageFilters.selection.environments,
-      ...normalizeDateTimeParams(pageFilters.selection.datetime),
-    };
-  }, [
-    environment,
-    pageFilters.selection.datetime,
-    pageFilters.selection.environments,
-    pageFilters.selection.projects,
-    project,
-    queryParams,
-  ]);
+  const {query, sort, viewId} = router.location.query;
 
   // This function is fired upon receiving new views from the backend - it replaces any previously
   // generated temporary view ids with the permanent view ids from the backend
@@ -417,7 +451,7 @@ export function IssueViewsPFStateProvider({
             normalizeUrl({
               ...location,
               query: {
-                ...queryParamsWithPageFilters,
+                ...router.location.query,
                 viewId: matchingView.id,
               },
             }),
@@ -435,13 +469,7 @@ export function IssueViewsPFStateProvider({
 
   const debounceUpdateViews = useMemo(
     () =>
-      debounce((newTabs: IssueViewPF[], pageFiltersSelection: PageFilters) => {
-        const isAllProjects =
-          pageFiltersSelection.projects.length === 1 &&
-          pageFiltersSelection.projects[0] === -1;
-
-        const projects = isAllProjects ? [] : pageFiltersSelection.projects;
-
+      debounce((newTabs: IssueViewPF[]) => {
         if (newTabs) {
           updateViews({
             orgSlug: organization.slug,
@@ -456,10 +484,10 @@ export function IssueViewsPFStateProvider({
                 name: tab.label,
                 query: tab.query,
                 querySort: tab.querySort,
-                projects,
-                isAllProjects,
-                environments: pageFiltersSelection.environments,
-                timeFilters: pageFiltersSelection.datetime,
+                projects: isEqual(tab.projects, [-1]) ? [] : tab.projects,
+                isAllProjects: isEqual(tab.projects, [-1]),
+                environments: tab.environments,
+                timeFilters: tab.timeFilters,
               })),
           });
         }
@@ -467,7 +495,7 @@ export function IssueViewsPFStateProvider({
     [organization.slug, updateViews]
   );
 
-  const reducer: Reducer<IssueViewsPFState, IssueViewsPFActions> = useCallback(
+  const reducer: Reducer<IssueViewsPFState, IssueViewsActions> = useCallback(
     (state, action): IssueViewsPFState => {
       if (!tabListState) {
         return state;
@@ -486,9 +514,9 @@ export function IssueViewsPFStateProvider({
         case 'DELETE_VIEW':
           return deleteView(state, tabListState);
         case 'CREATE_NEW_VIEW':
-          return createNewView(state, action);
+          return createNewView(state, action, defaultProject);
         case 'SET_TEMP_VIEW':
-          return setTempView(state, action);
+          return setTempView(state, action, defaultProject);
         case 'DISCARD_TEMP_VIEW':
           return discardTempView(state, tabListState);
         case 'SAVE_TEMP_VIEW':
@@ -505,7 +533,7 @@ export function IssueViewsPFStateProvider({
           return state;
       }
     },
-    [tabListState]
+    [tabListState, defaultProject]
   );
 
   const sortOption =
@@ -522,6 +550,9 @@ export function IssueViewsPFStateProvider({
           query: query.toString(),
           querySort: sortOption,
           isCommitted: true,
+          environments: DEFAULT_ENVIRONMENTS,
+          projects: defaultProject,
+          timeFilters: DEFAULT_TIME_FILTERS,
         }
       : undefined;
 
@@ -530,12 +561,12 @@ export function IssueViewsPFStateProvider({
     tempView: initialTempView,
   });
 
-  const dispatchWrapper = (action: IssueViewsPFActions) => {
+  const dispatchWrapper = (action: IssueViewsActions) => {
     const newState = reducer(state, action);
     dispatch(action);
 
     if (action.type === 'SYNC_VIEWS_TO_BACKEND' || action.syncViews) {
-      debounceUpdateViews(newState.views, pageFilters.selection);
+      debounceUpdateViews(newState.views);
     }
 
     const actionAnalyticsKey = ACTION_ANALYTICS_MAP[action.type];
@@ -565,7 +596,16 @@ export function IssueViewsPFStateProvider({
           querySort: IssueSortOptions.DATE,
           unsavedChanges: view.saveQueryToView
             ? undefined
-            : {query: view.query, querySort: IssueSortOptions.DATE},
+            : {
+                query: view.query,
+                querySort: IssueSortOptions.DATE,
+                environments: DEFAULT_ENVIRONMENTS,
+                projects: defaultProject,
+                timeFilters: DEFAULT_TIME_FILTERS,
+              },
+          environments: DEFAULT_ENVIRONMENTS,
+          projects: defaultProject,
+          timeFilters: DEFAULT_TIME_FILTERS,
           isCommitted: true,
         };
         return viewToTab;
@@ -579,8 +619,17 @@ export function IssueViewsPFStateProvider({
             querySort: IssueSortOptions.DATE,
             unsavedChanges: saveQueryToView
               ? undefined
-              : {query, querySort: IssueSortOptions.DATE},
+              : {
+                  query,
+                  querySort: IssueSortOptions.DATE,
+                  environments: DEFAULT_ENVIRONMENTS,
+                  projects: defaultProject,
+                  timeFilters: DEFAULT_TIME_FILTERS,
+                },
             isCommitted: true,
+            environments: DEFAULT_ENVIRONMENTS,
+            projects: defaultProject,
+            timeFilters: DEFAULT_TIME_FILTERS,
           };
         }
         return tab;
@@ -595,7 +644,7 @@ export function IssueViewsPFStateProvider({
         {
           ...location,
           query: {
-            ...queryParams,
+            ...router.location.query,
             query: newQuery,
             sort: IssueSortOptions.DATE,
           },
@@ -618,6 +667,7 @@ export function IssueViewsPFStateProvider({
         rootProps: {...restProps, orientation: 'horizontal'},
         tabListState,
         setTabListState,
+        defaultProject,
         dispatch: dispatchWrapper,
         state,
       }}
